memoized authenticator
this is NOT a permanent solution.

however, the auth.authenticate_shadow function was taking a
prohibitively long amount of time to auth each static request for a
file.

the realized downside of memoizing the authenticator is that if a user
changes the passwd (without restarting the mineos service), the previous
password will work IN ADDITION.

this might be solvable by putting a chokidar watch on the etc/shadows
password and re-memoizing, but the workableness of that solution is tbd
too.